Add explicit cast to place MSVC
git-svn-id: https://svn.osgeo.org/gdal/trunk@33494 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/gcore/overview.cpp
@@ -431,7 +431,7 @@ GDALResampleChunk32R_AverageT( double dfXRatioDstToSrc,
                         nB = (nTotalB + nCount / 2) / nCount;
                     pDstScanline[iDstPixel] = (T)GDALFindBestEntry(
                         nEntryCount, aEntries, nR, nG, nB,
-                        bHasNoData ? tNoDataValue : -1);
+                        bHasNoData ? static_cast<int>(tNoDataValue) : -1);
                 }
             }
         }
